# www/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Financial API",
        "description": "Wellcome! <br>To make requests, use login system passing cfpCnpj and password to get the token and click in Authorize and set the token",
        "contact": {
            "name": "Financial Team",
            "email": "gustamms@hotmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/type-transactions": {
            "get": {
                "tags": [
                    "Type transactions"
                ],
                "summary": "List of type transactions",
                "responses": {
                    "200": {
                        "description": "Returns data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "typ_tran_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "typ_tran_description": {
                                                        "type": "string",
                                                        "example": "Transferencia"
                                                    },
                                                    "typ_tran_sisid": {
                                                        "type": "string",
                                                        "example": "TRS"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2021-05-17T22:39:57.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2021-05-17T22:39:57.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "Unauthorized"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Type transactions"
                ],
                "summary": "Create new type transactions",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "typ_tran_description": {
                                        "type": "string",
                                        "example": "Transferencia"
                                    },
                                    "typ_tran_sisid": {
                                        "type": "string",
                                        "example": "TRS"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "typ_tran_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "typ_tran_description": {
                                                        "type": "string",
                                                        "example": "Transferencia"
                                                    },
                                                    "typ_tran_sisid": {
                                                        "type": "string",
                                                        "example": "TRS"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2021-05-17T22:39:57.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2021-05-17T22:39:57.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "Unauthorized"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/type-transactions/{id}": {
            "get": {
                "tags": [
                    "Type transactions"
                ],
                "summary": "Specific of type transactions",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Type transactions id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "typ_tran_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "typ_tran_description": {
                                                        "type": "string",
                                                        "example": "Transferencia"
                                                    },
                                                    "typ_tran_sisid": {
                                                        "type": "string",
                                                        "example": "TRS"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2021-05-17T22:39:57.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2021-05-17T22:39:57.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "Unauthorized"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with CPF/CNPJ and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}